按规范修改了下代码。提出book/effect作为公共的效果包
