+++ example/book-store/src/book/List.js
@@ -14,7 +14,6 @@ define(
             var book = this.model.find(e.isbn);
             var cart = require('cart/init');
             cart.add(book);
-
             this.view.showBoughtTip(e.isbn);
         }
 
@@ -29,11 +28,9 @@ define(
             var query = { page: e.page };
             var locator = this.model.get('locator');
             var URL = require('er/URL');
-            var cURL = URL.parse(location.hash);
-            locator.redirect(URL.withQuery(cURL.getPath(), require('er/util').mix(cURL.getQuery(),query)));
-            // 这个看看是否能在locator里给出这个直接获取当前的地址，或者说withQuery里给
-            //参数，比如cover = false 时除了增加后面的query，也保留原url的query。
-            // 这个在排序、翻页、筛选的时候会很方便
+            var cURL = this.model.get('url');
+            locator.redirect(URL.withQuery(cURL.getPath(),
+                require('er/util').mix(cURL.getQuery(), query)));
         }
 
         BookList.prototype.initBehavior = function() {
