+++ example/book-store/src/book/ListModel.js
@@ -24,7 +24,7 @@ define(
         BookListModel.prototype.prepare = function() {
             var list = this.get('list');
             this.set('list', list.result);
-            this.set('page', list.page ? list.page : 1);
+            this.set('page', list.page || 1);
             this.set('pageCount', list.pageCount);
 
             var pages = [];
